fix: refactor handlePrintToPDF method
